Moved linking policy to scripts, allowing per-script security.
